+++ src/routes/(auth)/setup/+page.server.ts
@@ -46,6 +46,6 @@ export const actions: Actions = {
       });
     }
 
-    throw redirect(303, "/movies");
+    throw redirect(303, "/continue");
   },
 };